Convert FRTB capital charges from DOUBLE PRECISION to NUMERIC
V10 alters all 8 capital charge columns in frtb_calculations from
DOUBLE PRECISION to NUMERIC(28,8) for exact reproducibility in
regulatory submissions. Updates Exposed table to decimal(28,8),
domain models and DTOs to BigDecimal, and all test assertions.

+++ regulatory-service/src/main/kotlin/com/kinetix/regulatory/model/FrtbCalculationRecord.kt
@@ -1,26 +1,27 @@
 package com.kinetix.regulatory.model
 
+import java.math.BigDecimal
 import java.time.Instant
 
 data class RiskClassCharge(
     val riskClass: String,
-    val deltaCharge: Double,
-    val vegaCharge: Double,
-    val curvatureCharge: Double,
-    val totalCharge: Double,
+    val deltaCharge: BigDecimal,
+    val vegaCharge: BigDecimal,
+    val curvatureCharge: BigDecimal,
+    val totalCharge: BigDecimal,
 )
 
 data class FrtbCalculationRecord(
     val id: String,
     val portfolioId: String,
-    val totalSbmCharge: Double,
-    val grossJtd: Double,
-    val hedgeBenefit: Double,
-    val netDrc: Double,
-    val exoticNotional: Double,
-    val otherNotional: Double,
-    val totalRrao: Double,
-    val totalCapitalCharge: Double,
+    val totalSbmCharge: BigDecimal,
+    val grossJtd: BigDecimal,
+    val hedgeBenefit: BigDecimal,
+    val netDrc: BigDecimal,
+    val exoticNotional: BigDecimal,
+    val otherNotional: BigDecimal,
+    val totalRrao: BigDecimal,
+    val totalCapitalCharge: BigDecimal,
     val sbmCharges: List<RiskClassCharge>,
     val calculatedAt: Instant,
     val storedAt: Instant,
